refactor(task): refactor 2.8.6 task

+++ tasks/2. Variables and arithmetic operations/2.8 Arithmetic operations +=, -=, mult=, div=, %=/Feat 06/main.c
@@ -9,7 +9,7 @@
 int main(void) {
   int price;
   int discount;
-  float total;
+  double total;
 
   scanf("%d %d", &price, &discount);
